Refactor: Pagination is now self made component
